strands-agents-sdk/pyproject.toml
[project]
name = "strands-agents-sdk"
version = "0.1.0"
description = "Examples demonstrating the Strands Agents SDK - a model-driven approach to building AI agents"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.29.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "strands-agents>=1.32.0",
    "strands-agents-tools>=0.2.23",
]
